店铺菜单管理

+++ src/main/java/com/kotall/rms/common/entity/litemall/LiteMallAdEntity.java
@@ -82,6 +82,6 @@ public class LiteMallAdEntity implements Serializable {
 	/**
 	 * 店铺ID
 	 */
-	private Integer storeId;
+	private Long storeId;
 
 }
