+++ src/main/java/com/kotall/rms/common/entity/litemall/LiteMallFootprintEntity.java
@@ -52,6 +52,6 @@ public class LiteMallFootprintEntity implements Serializable {
 	/**
 	 * 店铺ID
 	 */
-	private Integer storeId;
+	private Long storeId;
 	
 }
