+++ src/main/java/com/kotall/rms/common/entity/litemall/LiteMallOrderGoodsEntity.java
@@ -93,6 +93,6 @@ public class LiteMallOrderGoodsEntity implements Serializable {
 	/**
 	 * 店铺ID
 	 */
-	private Integer storeId;
+	private Long storeId;
 	
 }
