+++ src/main/java/com/kotall/rms/common/entity/litemall/LiteMallSearchHistoryEntity.java
@@ -57,5 +57,5 @@ public class LiteMallSearchHistoryEntity implements Serializable {
 	/**
 	 * 店铺ID
 	 */
-	private Integer storeId;
+	private Long storeId;
 }
